chore(api): move document review engine into lib/document-review
The review execution engine (ASK extraction, position grading, reference
comparison) now lives in lib so a background worker can call it without a
lib-to-handler import; handlers keep routing, schemas, permissions, and audit.
